Done some changes in project

+++ 006-EmployeeBackEnd/src/main/java/com/service/EmployeeServiceIMPL.java
@@ -1,8 +1,11 @@
 package com.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.dao.EmployeeRepository;
@@ -15,34 +18,63 @@ public class EmployeeServiceIMPL implements EmployeeService {
 	private EmployeeRepository employeeRepository;
 
 	@Override
-	public Employee saveEmployee(Employee emp) {
-		
-		return employeeRepository.save(emp);
+	public ResponseEntity<Employee> saveEmployee(Employee emp) {
+
+		if (employeeRepository.existsById(emp.getEid())) {
+			return new ResponseEntity<Employee>(employeeRepository.save(emp), HttpStatus.ALREADY_REPORTED);
+
+		} else {
+			return new ResponseEntity<Employee>(employeeRepository.save(emp), HttpStatus.CREATED);
+		}
 	}
 
 	@Override
-	public List<Employee> getAllEmployees() {
-		
-		return employeeRepository.findAll();
+	public ResponseEntity<List<Employee>> getAllEmployees() {
+
+		return new ResponseEntity<List<Employee>>(employeeRepository.findAll(), HttpStatus.OK);
 	}
 
 	@Override
-	public Employee getEmployeeById(int eid) {
+	public ResponseEntity<Employee> getEmployeeById(int eid) {
+
+		if (employeeRepository.existsById(eid)) {
+			return new ResponseEntity<Employee>(employeeRepository.findById(eid).get(), HttpStatus.ACCEPTED);
+
+		} else {
+			return new ResponseEntity<Employee>(new Employee(), HttpStatus.BAD_REQUEST);
+
+		}
+	}
+
+	@Override
+	public ResponseEntity<Employee> deleteEmployeeById(int eid) {
+		
+		  Optional<Employee> opt= employeeRepository.findById(eid);
+		
+		  if(opt.isPresent())
+		  {
+			  employeeRepository.deleteById(eid);
+			  return new ResponseEntity<Employee>(opt.get(), HttpStatus.ACCEPTED);
+		  }
+		  else
+		  {
+			  return new ResponseEntity<Employee>(new Employee(), HttpStatus.BAD_REQUEST);
+		  }
+
+	}
+
+	@Override
+	public ResponseEntity<Employee> updateEmployee(Employee emp) {
+
+		if(employeeRepository.existsById(emp.getEid()))
+		{
+			return new ResponseEntity<Employee>(employeeRepository.save(emp),HttpStatus.ACCEPTED);
+		}
+		else
+		{
+			return new ResponseEntity<Employee>(emp,HttpStatus.CONFLICT);
+		}
 	
-		return employeeRepository.findById(eid).get();
 	}
 
-	@Override
-	public Employee deleteEmployeeById(int eid) {
-		Employee emp = employeeRepository.findById(eid).get();
-		employeeRepository.deleteById(eid);
-		return emp;
-	}
-
-	@Override
-	public Employee updateEmployee(Employee emp) {
-		
-		return employeeRepository.save(emp);
-	}
-	
 }
